requirements.txt
equests>=2,<3
msal>=0,<2